update  notif

+++ app/Http/Controllers/ADMIN/NotificationController.php
@@ -61,7 +61,11 @@ class NotificationController extends Controller
 
         \App\Jobs\AssignGuideJob::dispatch($details);
 
-        Notification::where('booking_id', $booking->id)->delete();
+        $selectNotif = Notification::where('booking_id', $booking->id)->first();
+        if($selectNotif) {
+            Notification::where('booking_id', $booking->id)->delete();
+        }
+        
 
         $query = Notification::create([
             'booking_id' => $booking->id,
